+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2025-02-01
+
+### Added
+- Support for MoreMachine Advanced Factories (mekaf module):
+  - Liquefying Factory
+  - Dissolving Factory
+  - Oxidizing Factory
+  - Crystallizing Factory
+  - Washing Factory
+  - Centrifuging Factory
+  - Chemical Infusing Factory
+  - Pressurized Reacting Factory
+  - Solar Neutron Activating Factory
+- Separate config section for chemical factories (gas/fluid processing)
+
 ## [1.1.0] - 2025-02-01
 
 ### Added